+++ lib/rixie/agent/compressor.rb
@@ -23,8 +23,8 @@ module Rixie
         []
       end
 
-      def think(messages:, listener:)
-        internal_agent.think(messages:, listener:)
+      def think(messages:, listener:, schema: nil)
+        internal_agent.think(messages:, listener:, schema:)
       end
 
       private