+++ lib/rixie/strategy/re_act.rb
@@ -5,7 +5,7 @@ module Rixie
     class ReAct
       def run(task:, listener:)
         agent = Agent::ReAct.new(base_agent: task.agent)
-        run = Run.new(user_input: task.user_input, agent: agent, context: task.context)
+        run = Run.new(user_input: task.user_input, agent: agent, context: task.context, schema: task.schema)
         task.runs << run
         run.execute(listener:)
         run.output